# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FOODIE-FUND",
    "description": "眾資成城-在你心愛的餐廳成為合夥人"
  },
  "servers": [
    {
      "url": "https://foodiefund.onrender.com",
      "description": "Render"
    },
    {
      "url": "http://127.0.0.1:2330",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/v1/api/news": {
      "get": {
        "tags": [
          "公告管理"
        ],
        "summary": "公告清單查詢",
        "description": "公告清單查詢",
        "parameters": [
          {
            "name": "keyWord",
            "in": "query",
            "description": "關鍵字fuzzy[tittle,content], 預設空直為搜尋全部",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeSort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "清單顯示比數,default=10",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "顯示第幾頁資料default=1",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/api/news/admin": {
      "post": {
        "tags": [
          "公告管理"
        ],
        "summary": "新增公告",
        "description": "新增公告",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "description": "會員資料",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicAt": {
                    "example": "2024-09-03"
                  },
                  "title": {
                    "type": "string",
                    "example": "test"
                  },
                  "content": {
                    "type": "string",
                    "example": ""
                  },
                  "isTop": {
                    "type": "Boolean",
                    "default": false
                  },
                  "isEnabled": {
                    "type": "Boolean",
                    "default": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/api/news/admin/{id}": {
      "put": {
        "tags": [
          "公告管理"
        ],
        "summary": "更新基本資料",
        "description": "更新基本資料",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "required": true,
          "description": "會員資料",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "標題123"
                  },
                  "content": {
                    "type": "string",
                    "example": ""
                  },
                  "isTop": {
                    "type": "Boolean",
                    "default": false
                  },
                  "isEnabled": {
                    "type": "Boolean",
                    "default": true
                  },
                  "publicAt": {
                    "type": "Date",
                    "example": ""
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "公告管理"
        ],
        "summary": "刪除單筆公告",
        "description": "刪除單筆公告",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者id"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}